Client ID is 16 instead of 32 bit

+++ Hangman-server/Player.cpp
@@ -7,7 +7,7 @@
 
 #include "Player.hpp"
 
-Player::Player(uint32_t restorationId) {
+Player::Player(uint16_t restorationId) {
     this->restorationId = restorationId;
 }
